Move dynamic image & text controls to packages directory
